refactor: group decode-reuse helpers in reuse.go
Move sliceHeader, noPointers, reuseOrMakeSlice (out of the 1466-line
reflect_encode.go) and reuseOrMakeMap (out of the 1-function maps_reuse.go)
into a single cohesive reuse.go — the decode-time "reuse the caller's
container" cluster. Deletes the 1-function file; no behavior change.

Also fix TestMapReuseCutsAllocs to measure a DIRECT struct-with-map target
(where map-reuse actually fires) instead of a []struct batch (whose maps the
slice-reuse element-clear nils before the map decoder runs, so it showed ~0).
